feat(Tests): reduce the number of stubs by using anonymous classes
